+++ source/RewardsService/rewardsservice/handlers/rewards_base_handler.py
